android: make the code be compatible with stlport
The android's stlport doesn't have tr1/unordered_set but unordered_set.
However, there are some unexpected errors on instantiating a user-defined
type in the unordered_set template. The errors are so strange that I can't
understand why. Probably a bug of stlport or g++ itself. However,
instantiating a built-in type is fine.

So instead of using unordered_set<user-defined-type *>, the patch changes
to use unordered_set<void *> and adds a wrapper template class to convert
the iterators to the expected user-defined type.

Signed-off-by: Chih-Wei Huang <cwhuang@linux.org.tw>